[FEATURE] Add cypress custom commands to support better type events #364
